Fix Id of histogram related entities and repositories

+++ server/HubMap/src/main/java/br/com/pucminas/hubmap/domain/indexing/HistogramItemRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(exported = false)
-public interface HistogramItemRepository extends CrudRepository<HistogramItem, Integer>{
+public interface HistogramItemRepository extends CrudRepository<HistogramItem, Long>{
 
 }
