+++ server/HubMap/src/main/java/br/com/pucminas/hubmap/domain/indexing/HistogramRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(exported = false)
-public interface HistogramRepository extends CrudRepository<Histogram, Integer>{
+public interface HistogramRepository extends CrudRepository<Histogram, Long>{
 
 }
